📝 feat.검색페이지 컴포 폴더명 및 경로변경
