+++ src/components/searchPage/searchNav/SearchNav.style.ts
@@ -1,24 +1,13 @@
 import styled from 'styled-components'
 
-const SearchNav = () => {
-  return (
-    <>
-      <LineDiv>
-        <span>검색결과</span>
-      </LineDiv>
-    </>
-  )
-}
-export default SearchNav
-
-const LineDiv = styled.div`
-  display: flex;
+export const LineDiv = styled.div`
   width: 97%;
+  display: flex;
+  padding-bottom: 15px;
+  margin: 0px auto;
   border-width: 0 0 1px;
   border-style: solid;
   border-color: ${({ theme }) => theme.rgba04};
-  margin: 0px auto;
-  padding-bottom: 15px;
   & span {
     font-size: ${({ theme }) => theme.fs15};
     color: ${({ theme }) => theme.fc09};
