+++ src/components/searchPage/searchNav/SearchNav.tsx
@@ -1,0 +1,12 @@
+import * as t from './SearchNav.style'
+
+const SearchNav = () => {
+  return (
+    <>
+      <t.LineDiv>
+        <span>검색결과</span>
+      </t.LineDiv>
+    </>
+  )
+}
+export default SearchNav
